% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_incidence_rate.R
\name{as_incidence_rate}
\alias{as_incidence_rate}
\title{as_incidence_rate}
\usage{
as_incidence_rate(x, lower, upper, units = "person-1 yr-1")
}
\arguments{
\item{x}{(numeric)}

\item{units}{(character) default is "per million" persons per year}
}
\value{
units
}
\description{
Standardize incidence rate
}
