% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PointRange.R
\name{est}
\alias{est}
\title{----------------------------------------------------------------------}
\usage{
est(x)
}
\description{
Accessors: `est()`, `lower()` and `upper()`
}
\details{
----------------------------------------------------------------------
}
