% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general-functions.R
\name{expand_codes}
\alias{expand_codes}
\title{plot_grid: Flexible function to plot data from get_data}
\usage{
expand_codes(codes)
}
\arguments{
\item{codes}{Data frame from get_data}
}
\description{
plot_grid: Flexible function to plot data from get_data
}
\keyword{plot}
