% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_WONDER_UCD.R
\name{import_WONDER_UCD}
\alias{import_WONDER_UCD}
\title{import_WONDER_UCD}
\usage{
import_WONDER_UCD(
  path,
  na = c("", "Suppressed", "Not Applicable", "Unreliable"),
  ...
)
}
\arguments{
\item{path}{(character) path to .txt file}

\item{na}{values to replace with NA}

\item{...reserved}{for future use}
}
\value{
tibble
}
\description{
import_WONDER_UCD
}
