% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PointRange.R
\name{pillar_shaft.vctrs_ptrg}
\alias{pillar_shaft.vctrs_ptrg}
\title{pillar_shaft}
\usage{
\method{pillar_shaft}{vctrs_ptrg}(x, ..., trim = FALSE)
}
\arguments{
\item{x}{`ptrg` object}

\item{...}{further arguments to `format()`}
}
\description{
pillar_shaft
}
