% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line-graphs.R
\name{plot_grid}
\alias{plot_grid}
\title{plot_grid: Flexible function to plot data from get_data}
\usage{
plot_grid(
  df,
  minwonderyear = 1999,
  maxwonderyear = 2018,
  groups = c(sprintf("No Expansion Before \%s", maxwonderyear), "Expanded in 2014",
    sprintf("Expanded Between 2015 and \%s", maxwonderyear)),
  group_conditions = c("is.na(aca_date) | year(aca_date) > maxwonderyear",
    "year(aca_date) == 2014", "T"),
  grid_vars = c("sex", "race"),
  ROOTPATH = "",
  RUNNAME = "year_state_sex_race_firearms",
  save = T,
  out_fn = T,
  listwise = T,
  vline = T,
  include_n = F,
  width = 6,
  height = 4,
  xlab = "Year",
  ylab = "Crude Rate\\n(Per 100k Deaths)",
  group_title = "",
  colpalette = "Dark2"
)
}
\arguments{
\item{df}{Data frame from get_data}

\item{minwonderyear}{minimum year from Wonder, default 1999}

\item{maxwonderyear}{maximum year from wonder, default 2018}

\item{groups}{These are for multiple lines in the same plot. Name the groups here....}

\item{group_conditions}{... and set their conditions as a string here.}

\item{grid_vars}{These are for multiple plots in a grid}

\item{ROOTPATH}{ROOTPATH that includes input and output folders}

\item{RUNNAME}{Name for the data}

\item{save}{Default True, if False does not save the plot}

\item{out_fn}{Default RUNNAME.png}

\item{listwise}{Default False, if True then delete all groups that don't have the max number of observations}

\item{vline}{If numeric, locaiton of a vertical line}

\item{include_n}{Defualt False, if True then group labels are given Ns}

\item{width}{Width of the plot in inches}

\item{height}{Height of the plot in inches}

\item{xlab}{X axis label, Default Year}

\item{ylab}{Y axis label, default Crude Rate Per 100,000}

\item{group_title}{Name of groups, default "",}

\item{colpalette}{ggplot2 colortheme}
}
\description{
plot_grid: Flexible function to plot data from get_data
}
\examples{
plot_grid(
  df,
  minwonderyear = 1999,
  maxwonderyear = 2018,
  groups = NULL,
  group_conditions = NULL,
  grid_vars = c("race"),
  ROOTPATH = "C:/Users/tcapu/Desktop/wonderdd",
  RUNNAME = "year_state_sex_race_opioids",
  listwise = F,
  vline = F,
  save = T,
  width = 6,
  height = 4,
  group_title="",
  colpalette="Dark2"
)
}
\keyword{plot}
