% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PointRange.R
\name{name ptrg-units}
\alias{name ptrg-units}
\alias{drop_units.vctrs_ptrg}
\alias{set_units.vctrs_ptrg}
\alias{get_units.vctrs_ptrg}
\title{----------------------------------------------------------------------}
\usage{
\method{drop_units}{vctrs_ptrg}(x)

\method{set_units}{vctrs_ptrg}(x, ...)

\method{get_units}{vctrs_ptrg}(x)
}
\arguments{
\item{x}{`ptrg` object}

\item{...}{(optional)}
}
\description{
Unit-awareness
}
\details{
----------------------------------------------------------------------
Unit-aware methods
}
