% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PointRange.R
\name{ptrg}
\alias{ptrg}
\title{----------------------------------------------------------------------}
\usage{
ptrg(est = double(), lower = double(), upper = double())
}
\arguments{
\item{est}{estimate}

\item{lower}{lower bound}

\item{upper}{upper bound}
}
\description{
Constructors
}
\details{
----------------------------------------------------------------------
ptrg

Constructor for new `ptrg` objects
}
