% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{statevars}
\alias{statevars}
\title{statevars}
\format{
Year by state dataset with several variables
}
\source{

}
\usage{
data(statevars)
}
\description{
statevars
}
\examples{
statevars
}
\keyword{dataset}
