% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PointRange.R
\name{vec_cast.vctrs_ptrg.vctrs_ptrg}
\alias{vec_cast.vctrs_ptrg.vctrs_ptrg}
\title{----------------------------------------------------------------------}
\usage{
\method{vec_cast}{vctrs_ptrg.vctrs_ptrg}(x, to, ...)
}
\description{
Coercion: `vec_cast()` and `vec_ptype2()`
}
\details{
----------------------------------------------------------------------
}
