% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PointRange.R
\name{vec_ptype_abbr.vctrs_ptrg}
\alias{vec_ptype_abbr.vctrs_ptrg}
\title{----------------------------------------------------------------------}
\usage{
\method{vec_ptype_abbr}{vctrs_ptrg}(x)
}
\description{
Formatting methods
}
\details{
- `vec_ptype_abbr`
- `obj_print_header`
- `format`
- `pillar_shaft`
- `type_sum`
- etc.

----------------------------------------------------------------------
}
